Add support for choosing next/prev subtitle lines using g/y keys.
Mplayer have used g/y keys to choose prev/next subtitle line
for a long time. MPlayer OSX Extended have the same ability
up to rev12 and then the functionality disappeared. This
patch restores the ability to use g/y because it is a very
useful way to sync subtitles that are a little bit off.

Signed-Off-By: Georgi Chorbadzhiyski <georgi@unixsol.org>